+++ raphtory/src/python/graph/edge.rs
@@ -14,21 +14,10 @@ use crate::{
                 StaticGraphViewOps,
             },
         },
-        graph::{edge::EdgeView, edges::NestedEdges, views::deletion_graph::GraphWithDeletions},
+        graph::{edge::EdgeView, views::deletion_graph::GraphWithDeletions},
     },
     prelude::*,
-    python::{
-        graph::properties::PyNestedPropsIterable,
-        types::{
-            repr::{iterator_repr, Repr},
-            wrappers::iterators::{
-                NestedArcStringVecIterable, NestedBoolIterable, NestedI64VecIterable,
-                NestedOptionArcStringIterable, NestedOptionI64Iterable, NestedU64U64Iterable,
-                NestedUtcDateTimeIterable, NestedVecUtcDateTimeIterable,
-            },
-        },
-        utils::PyTime,
-    },
+    python::{types::repr::Repr, utils::PyTime},
 };
 use chrono::{DateTime, Utc};
 use itertools::Itertools;
@@ -136,18 +125,6 @@ impl IntoPy<PyObject> for EdgeView<MaterializedGraph, MaterializedGraph> {
         Py::new(py, (PyMutableEdge::from(self.clone()), PyEdge::from(self)))
             .unwrap() // I think this only fails if we are out of memory? Seems to be unavoidable!
             .into_py(py)
-    }
-}
-
-impl IntoPy<PyObject> for ArcStr {
-    fn into_py(self, py: Python<'_>) -> PyObject {
-        self.0.into_py(py)
-    }
-}
-
-impl<'source> FromPyObject<'source> for ArcStr {
-    fn extract(ob: &'source PyAny) -> PyResult<Self> {
-        ob.extract::<String>().map(|v| v.into())
     }
 }
 impl_edgeviewops!(PyEdge, edge, EdgeView<DynamicGraph>, "Edge");
@@ -378,169 +355,6 @@ impl PyMutableEdge {
 
     fn __repr__(&self) -> String {
         self.repr()
-    }
-}
-
-#[pyclass(name = "NestedEdges")]
-pub struct PyNestedEdges {
-    edges: NestedEdges<'static, DynamicGraph>,
-}
-
-impl_edgeviewops!(
-    PyNestedEdges,
-    edges,
-    NestedEdges<'static, DynamicGraph>,
-    "NestedEdges"
-);
-impl_iterable_mixin!(
-    PyNestedEdges,
-    edges,
-    Vec<Vec<EdgeView<DynamicGraph>>>,
-    "list[list[Edges]]",
-    "edge"
-);
-
-impl<G: StaticGraphViewOps + IntoDynamic, GH: StaticGraphViewOps + IntoDynamic> IntoPy<PyObject>
-    for NestedEdges<'static, G, GH>
-{
-    fn into_py(self, py: Python<'_>) -> PyObject {
-        let edges = NestedEdges {
-            graph: self.graph.into_dynamic(),
-            nodes: self.nodes,
-            base_graph: self.base_graph.into_dynamic(),
-            edges: self.edges,
-        };
-        PyNestedEdges { edges }.into_py(py)
-    }
-}
-
-impl<'graph, G: GraphViewOps<'graph>, GH: GraphViewOps<'graph>> Repr
-    for NestedEdges<'graph, G, GH>
-{
-    fn repr(&self) -> String {
-        format!("NestedEdges({})", iterator_repr(self.iter()))
-    }
-}
-#[pymethods]
-impl PyNestedEdges {
-    /// Returns the earliest time of the edges.
-    #[getter]
-    fn earliest_time(&self) -> NestedOptionI64Iterable {
-        let edges = self.edges.clone();
-        (move || edges.earliest_time()).into()
-    }
-
-    /// Returns the earliest date time of the edges.
-    #[getter]
-    fn earliest_date_time(&self) -> NestedUtcDateTimeIterable {
-        let edges = self.edges.clone();
-        (move || edges.earliest_date_time()).into()
-    }
-
-    /// Returns the latest time of the edges.
-    #[getter]
-    fn latest_time(&self) -> NestedOptionI64Iterable {
-        let edges = self.edges.clone();
-        (move || edges.latest_time()).into()
-    }
-
-    /// Returns the latest date time of the edges.
-    #[getter]
-    fn latest_date_time(&self) -> NestedUtcDateTimeIterable {
-        let edges = self.edges.clone();
-        (move || edges.latest_date_time()).into()
-    }
-
-    /// Returns the times of exploded edges
-    #[getter]
-    fn time(&self) -> NestedOptionI64Iterable {
-        let edges = self.edges.clone();
-        (move || edges.time()).into()
-    }
-
-    /// Returns the name of the layer the edges belong to - assuming they only belong to one layer
-    #[getter]
-    fn layer_name(&self) -> NestedOptionArcStringIterable {
-        let edges = self.edges.clone();
-        (move || edges.layer_name()).into()
-    }
-
-    /// Returns the names of the layers the edges belong to
-    #[getter]
-    fn layer_names(&self) -> NestedArcStringVecIterable {
-        let edges = self.edges.clone();
-        (move || {
-            edges.layer_names().map(
-                |e: Box<dyn Iterator<Item = Box<dyn Iterator<Item = ArcStr> + Send>> + Send>| {
-                    e.map(|e| e.collect_vec())
-                },
-            )
-        })
-        .into()
-    }
-
-    // FIXME: needs a view that allows indexing into the properties
-    /// Returns all properties of the edges
-    #[getter]
-    fn properties(&self) -> PyNestedPropsIterable {
-        let edges = self.edges.clone();
-        (move || edges.properties()).into()
-    }
-
-    /// Returns all ids of the edges.
-    #[getter]
-    fn id(&self) -> NestedU64U64Iterable {
-        let edges = self.edges.clone();
-        (move || edges.id()).into()
-    }
-
-    /// Returns all timestamps of edges, when an edge is added or change to an edge is made.
-    fn history(&self) -> NestedI64VecIterable {
-        let edges = self.edges.clone();
-        (move || edges.history()).into()
-    }
-
-    /// Returns all timestamps of edges, when an edge is added or change to an edge is made.
-    fn history_date_time(&self) -> NestedVecUtcDateTimeIterable {
-        let edges = self.edges.clone();
-        (move || edges.history_date_time()).into()
-    }
-
-    /// Returns all timestamps of edges, where an edge is deleted
-    ///
-    /// Returns:
-    ///     A list of lists of lists of unix timestamps
-    fn deletions(&self) -> NestedI64VecIterable {
-        let edges = self.edges.clone();
-        (move || edges.deletions()).into()
-    }
-
-    /// Returns all timestamps of edges, where an edge is deleted
-    ///
-    /// Returns:
-    ///     A list of lists of lists of DateTime objects
-    fn deletions_date_time(&self) -> NestedVecUtcDateTimeIterable {
-        let edges = self.edges.clone();
-        (move || edges.deletions_date_time()).into()
-    }
-
-    /// Check if edges are valid (i.e., not deleted)
-    fn is_valid(&self) -> NestedBoolIterable {
-        let edges = self.edges.clone();
-        (move || edges.is_valid()).into()
-    }
-
-    /// Check if edges are deleted
-    fn is_deleted(&self) -> NestedBoolIterable {
-        let edges = self.edges.clone();
-        (move || edges.is_deleted()).into()
-    }
-
-    /// Get the date times of exploded edges
-    #[getter]
-    fn date_time(&self) -> NestedUtcDateTimeIterable {
-        let edges = self.edges.clone();
-        (move || edges.date_time()).into()
     }
 }
