+++ raphtory/src/python/types/mod.rs
@@ -1,6 +1,7 @@
 #[macro_use]
 pub mod macros;
 
+mod arcstr;
 pub mod iterable;
 pub mod repr;
 pub mod wrappers;
